+++ main/main.go
@@ -21,14 +21,6 @@ import (
 	"github.com/stackrox/k8s-istio-cve-pusher/utils"
 )
 
-const (
-	cpeVendorIndex  = 3
-	cpeProductIndex = 4
-	cpeVersionIndex = 5
-	cpeUpdateIndex  = 6
-	cpePartsLength  = 13
-)
-
 var (
 	// specs is the set of projects that we wish to filter CVEs for.
 	specs = map[nvd.Project]string{
@@ -67,8 +59,8 @@ func runCmd() error {
 	groups := splitCVEs(specs, allCVEs)
 
 	// Update fixed by version for NVD CVEs that do not have it but has been determined by manual audit.
-	for project, cves := range groups {
-		updateFixedByVersionIfMissing(project, cves...)
+	for _, cves := range groups {
+		nvd.FillMissingData(cves...)
 	}
 
 	// Do some logging for the resulting groups of CVEs.
