+++ main/main_test.go
@@ -4,7 +4,6 @@ import (
 	"io/ioutil"
 	"testing"
 
-	"github.com/facebookincubator/nvdtools/cvefeed/nvd/schema"
 	"github.com/opentracing/opentracing-go/log"
 	"github.com/stackrox/k8s-istio-cve-pusher/nvd"
 	"github.com/stretchr/testify/assert"
@@ -78,120 +77,4 @@ func TestChecksum(t *testing.T) {
 	require.Nil(t, err)
 	assert.True(t, len(data) > 0)
 	assert.Equal(t, "8a8c5bce18d9640a5bac6e0ec553c5d67ae25645a9f580fea2f6362ef19ff0e7", checksum)
-}
-
-func TestUpdateFixedByWithNoDataFromNVD(t *testing.T) {
-	cve := &schema.NVDCVEFeedJSON10DefCVEItem{
-		CVE: &schema.CVEJSON40{
-			CVEDataMeta: &schema.CVEJSON40CVEDataMeta{
-				ID: "CVE-2016-7075",
-			},
-		},
-		Configurations: &schema.NVDCVEFeedJSON10DefConfigurations{
-			Nodes: []*schema.NVDCVEFeedJSON10DefNode{
-				{
-					Operator: "OR",
-					CPEMatch: []*schema.NVDCVEFeedJSON10DefCPEMatch{
-						{
-							Vulnerable: true,
-							Cpe23Uri:   "cpe:2.3:a:kubernetes:kubernetes:-:*:*:*:*:*:*:*",
-						},
-					},
-				},
-			},
-		},
-	}
-	expectedCPENodes := []*schema.NVDCVEFeedJSON10DefNode{
-		{
-			Operator: "OR",
-			CPEMatch: []*schema.NVDCVEFeedJSON10DefCPEMatch{
-				{
-					Vulnerable: true,
-					Cpe23Uri:   "cpe:2.3:a:kubernetes:kubernetes:-:*:*:*:*:*:*:*",
-				},
-			},
-		},
-		{
-			Operator: "OR",
-			CPEMatch: []*schema.NVDCVEFeedJSON10DefCPEMatch{
-				{
-					Cpe23Uri:            "cpe:2.3:a:kubernetes:kubernetes:*:*:*:*:*:*:*:*",
-					VersionEndExcluding: "1.2.7",
-				},
-				{
-					Cpe23Uri:              "cpe:2.3:a:kubernetes:kubernetes:*:*:*:*:*:*:*:*",
-					VersionStartIncluding: "1.3.0",
-					VersionEndExcluding:   "1.3.9",
-				},
-				{
-					Cpe23Uri:              "cpe:2.3:a:kubernetes:kubernetes:*:*:*:*:*:*:*:*",
-					VersionStartIncluding: "1.4.0",
-					VersionEndExcluding:   "1.4.3",
-				},
-			},
-		},
-	}
-
-	updateFixedByVersionIfMissing(nvd.Kubernetes, cve)
-
-	assert.EqualValues(t, expectedCPENodes, cve.Configurations.Nodes)
-
-}
-
-func TestUpdateFixedByWithDataFromNVD(t *testing.T) {
-	cve := &schema.NVDCVEFeedJSON10DefCVEItem{
-		CVE: &schema.CVEJSON40{
-			CVEDataMeta: &schema.CVEJSON40CVEDataMeta{
-				ID: "CVE-2016-7075",
-			},
-		},
-		Configurations: &schema.NVDCVEFeedJSON10DefConfigurations{
-			Nodes: []*schema.NVDCVEFeedJSON10DefNode{
-				{
-					Operator: "OR",
-					CPEMatch: []*schema.NVDCVEFeedJSON10DefCPEMatch{
-						{
-							Cpe23Uri:            "cpe:2.3:a:kubernetes:kubernetes:*:*:*:*:*:*:*:*",
-							VersionEndExcluding: "1.2.7",
-						},
-						{
-							Cpe23Uri:              "cpe:2.3:a:kubernetes:kubernetes:*:*:*:*:*:*:*:*",
-							VersionStartIncluding: "1.3.0",
-							VersionEndExcluding:   "1.3.9",
-						},
-						{
-							Cpe23Uri:              "cpe:2.3:a:kubernetes:kubernetes:*:*:*:*:*:*:*:*",
-							VersionStartIncluding: "1.4.0",
-							VersionEndExcluding:   "1.4.3",
-						},
-					},
-				},
-			},
-		},
-	}
-	expectedCPENodes := []*schema.NVDCVEFeedJSON10DefNode{
-		{
-			Operator: "OR",
-			CPEMatch: []*schema.NVDCVEFeedJSON10DefCPEMatch{
-				{
-					Cpe23Uri:            "cpe:2.3:a:kubernetes:kubernetes:*:*:*:*:*:*:*:*",
-					VersionEndExcluding: "1.2.7",
-				},
-				{
-					Cpe23Uri:              "cpe:2.3:a:kubernetes:kubernetes:*:*:*:*:*:*:*:*",
-					VersionStartIncluding: "1.3.0",
-					VersionEndExcluding:   "1.3.9",
-				},
-				{
-					Cpe23Uri:              "cpe:2.3:a:kubernetes:kubernetes:*:*:*:*:*:*:*:*",
-					VersionStartIncluding: "1.4.0",
-					VersionEndExcluding:   "1.4.3",
-				},
-			},
-		},
-	}
-
-	updateFixedByVersionIfMissing(nvd.Kubernetes, cve)
-
-	assert.EqualValues(t, expectedCPENodes, cve.Configurations.Nodes)
 }
