+++ nvd/cve.go
@@ -1,4 +1,4 @@
-package main
+package nvd
 
 import (
 	"strings"
@@ -6,7 +6,17 @@ import (
 	"github.com/facebookincubator/nvdtools/cvefeed/nvd/schema"
 	"github.com/pkg/errors"
 	log "github.com/sirupsen/logrus"
-	"github.com/stackrox/k8s-istio-cve-pusher/nvd"
+)
+
+const (
+	cpeVersionIndex = 5
+	cpeUpdateIndex  = 6
+	cpePartsLength  = 13
+
+	// ANY represent CPE ANY string punctuation
+	ANY = "*"
+	// NOT_APPLICABLE represents N/A CPE string punctuation
+	NOT_APPLICABLE = "-"
 )
 
 var (
@@ -64,7 +74,11 @@ var (
 	}
 )
 
-func updateFixedByVersionIfMissing(project nvd.Project, cves ...*schema.NVDCVEFeedJSON10DefCVEItem) {
+func FillMissingData(cves ...*schema.NVDCVEFeedJSON10DefCVEItem) {
+	tryUpdateCVEWithAffectedVersions(cves...)
+}
+
+func tryUpdateCVEWithAffectedVersions(cves ...*schema.NVDCVEFeedJSON10DefCVEItem) {
 	for _, cve := range cves {
 		if _, ok := cvesWithoutFixedBy[cve.CVE.CVEDataMeta.ID]; !ok {
 			continue
@@ -73,27 +87,14 @@ func updateFixedByVersionIfMissing(project nvd.Project, cves ...*schema.NVDCVEFe
 		updateRequired := true
 		for _, node := range cve.Configurations.Nodes {
 			for _, cpeMatch := range node.CPEMatch {
-				cpeVersionAndUpdate, err := getVersionAndUpdateFromCPE(cpeMatch.Cpe23Uri, project)
+				hasVersionInfo, err := cpeHasVersionInfo(cpeMatch)
 				if err != nil {
-					log.Error(errors.Wrapf(err, "could not get version and update from cpe: %q", cpeMatch.Cpe23Uri))
+					log.Error(err)
 					continue
 				}
 
-				// Possibly invalid CPE or non-kube and non-istio CPE.
-				if cpeVersionAndUpdate == "" {
-					continue
-				}
-
-				// We need to update CVE only if none of the CPE provide any information about affected version
-				if cpeVersionAndUpdate != "-:*" {
-					// If affected version is specified move to next CPE
-					if cpeVersionAndUpdate != "*:*" {
-						continue
-					}
-					// If CPE string does not specify affected version, look for version range
-					if cpeMatch.VersionStartIncluding != "" || cpeMatch.VersionEndIncluding != "" || cpeMatch.VersionEndExcluding != "" {
-						updateRequired = false
-					}
+				if hasVersionInfo {
+					updateRequired = false
 				}
 			}
 		}
@@ -105,26 +106,38 @@ func updateFixedByVersionIfMissing(project nvd.Project, cves ...*schema.NVDCVEFe
 	}
 }
 
-func getVersionAndUpdateFromCPE(cpe string, project nvd.Project) (string, error) {
+func getVersionAndUpdate(cpe string) (string, string, error) {
 	if ok := strings.HasPrefix(cpe, "cpe:2.3:a:"); !ok {
-		return "", errors.Errorf("cpe: %q not a valid cpe23Uri format", cpe)
+		return "", "", errors.Errorf("cpe: %q not a valid cpe23Uri format", cpe)
 	}
 
 	ss := strings.Split(cpe, ":")
 	if len(ss) != cpePartsLength {
-		return "", errors.Errorf("cpe: %q not a valid cpe23Uri format", cpe)
+		return "", "", errors.Errorf("cpe: %q not a valid cpe23Uri format", cpe)
 	}
-	if _, ok := specs[project]; !ok {
-		return "", errors.Errorf("unknown CVE type: %s", project.String())
+	return ss[cpeVersionIndex], ss[cpeUpdateIndex], nil
+}
+
+func cpeHasVersionInfo(cpeMatch *schema.NVDCVEFeedJSON10DefCPEMatch) (bool, error) {
+	version, update, err := getVersionAndUpdate(cpeMatch.Cpe23Uri)
+	if err != nil {
+		return false, errors.Wrapf(err, "could not get version and update from cpe: %q", cpeMatch.Cpe23Uri)
 	}
-	if project == nvd.Kubernetes && (ss[cpeVendorIndex] != "kubernetes" || ss[cpeProductIndex] != "kubernetes") {
-		return "", nil
+
+	// Possibly invalid CPE
+	if version == "" && update == "" {
+		return false, nil
 	}
-	if project == nvd.Istio && (ss[cpeVendorIndex] != "istio" || ss[cpeProductIndex] != "istio") {
-		return "", nil
+
+	// CPE string does not provide affected versions information
+	if version == NOT_APPLICABLE {
+		return false, nil
 	}
-	if project == nvd.Openshift && (ss[cpeVendorIndex] != "redhat" || ss[cpeProductIndex] != "openshift_container_platform") {
-		return "", nil
+
+	// If CPE string matches any version check if ranges are provided. Note: CPE23URI provides affected versions and not ranges.
+	if version == ANY && update == ANY {
+		return cpeMatch.VersionStartIncluding != "" || cpeMatch.VersionEndIncluding != "" || cpeMatch.VersionEndExcluding != "", nil
 	}
-	return strings.Join(ss[cpeVersionIndex:cpeUpdateIndex+1], ":"), nil
+
+	return true, nil
 }
