+++ nvd/cve_test.go
@@ -1,0 +1,124 @@
+package nvd
+
+import (
+	"testing"
+
+	"github.com/facebookincubator/nvdtools/cvefeed/nvd/schema"
+	"github.com/stretchr/testify/assert"
+)
+
+func TestUpdateFixedByWithNoDataFromNVD(t *testing.T) {
+	cve := &schema.NVDCVEFeedJSON10DefCVEItem{
+		CVE: &schema.CVEJSON40{
+			CVEDataMeta: &schema.CVEJSON40CVEDataMeta{
+				ID: "CVE-2016-7075",
+			},
+		},
+		Configurations: &schema.NVDCVEFeedJSON10DefConfigurations{
+			Nodes: []*schema.NVDCVEFeedJSON10DefNode{
+				{
+					Operator: "OR",
+					CPEMatch: []*schema.NVDCVEFeedJSON10DefCPEMatch{
+						{
+							Vulnerable: true,
+							Cpe23Uri:   "cpe:2.3:a:kubernetes:kubernetes:-:*:*:*:*:*:*:*",
+						},
+					},
+				},
+			},
+		},
+	}
+	expectedCPENodes := []*schema.NVDCVEFeedJSON10DefNode{
+		{
+			Operator: "OR",
+			CPEMatch: []*schema.NVDCVEFeedJSON10DefCPEMatch{
+				{
+					Vulnerable: true,
+					Cpe23Uri:   "cpe:2.3:a:kubernetes:kubernetes:-:*:*:*:*:*:*:*",
+				},
+			},
+		},
+		{
+			Operator: "OR",
+			CPEMatch: []*schema.NVDCVEFeedJSON10DefCPEMatch{
+				{
+					Cpe23Uri:            "cpe:2.3:a:kubernetes:kubernetes:*:*:*:*:*:*:*:*",
+					VersionEndExcluding: "1.2.7",
+				},
+				{
+					Cpe23Uri:              "cpe:2.3:a:kubernetes:kubernetes:*:*:*:*:*:*:*:*",
+					VersionStartIncluding: "1.3.0",
+					VersionEndExcluding:   "1.3.9",
+				},
+				{
+					Cpe23Uri:              "cpe:2.3:a:kubernetes:kubernetes:*:*:*:*:*:*:*:*",
+					VersionStartIncluding: "1.4.0",
+					VersionEndExcluding:   "1.4.3",
+				},
+			},
+		},
+	}
+
+	tryUpdateCVEWithAffectedVersions(cve)
+
+	assert.EqualValues(t, expectedCPENodes, cve.Configurations.Nodes)
+
+}
+
+func TestUpdateFixedByWithDataFromNVD(t *testing.T) {
+	cve := &schema.NVDCVEFeedJSON10DefCVEItem{
+		CVE: &schema.CVEJSON40{
+			CVEDataMeta: &schema.CVEJSON40CVEDataMeta{
+				ID: "CVE-2016-7075",
+			},
+		},
+		Configurations: &schema.NVDCVEFeedJSON10DefConfigurations{
+			Nodes: []*schema.NVDCVEFeedJSON10DefNode{
+				{
+					Operator: "OR",
+					CPEMatch: []*schema.NVDCVEFeedJSON10DefCPEMatch{
+						{
+							Cpe23Uri:            "cpe:2.3:a:kubernetes:kubernetes:*:*:*:*:*:*:*:*",
+							VersionEndExcluding: "1.2.7",
+						},
+						{
+							Cpe23Uri:              "cpe:2.3:a:kubernetes:kubernetes:*:*:*:*:*:*:*:*",
+							VersionStartIncluding: "1.3.0",
+							VersionEndExcluding:   "1.3.9",
+						},
+						{
+							Cpe23Uri:              "cpe:2.3:a:kubernetes:kubernetes:*:*:*:*:*:*:*:*",
+							VersionStartIncluding: "1.4.0",
+							VersionEndExcluding:   "1.4.3",
+						},
+					},
+				},
+			},
+		},
+	}
+	expectedCPENodes := []*schema.NVDCVEFeedJSON10DefNode{
+		{
+			Operator: "OR",
+			CPEMatch: []*schema.NVDCVEFeedJSON10DefCPEMatch{
+				{
+					Cpe23Uri:            "cpe:2.3:a:kubernetes:kubernetes:*:*:*:*:*:*:*:*",
+					VersionEndExcluding: "1.2.7",
+				},
+				{
+					Cpe23Uri:              "cpe:2.3:a:kubernetes:kubernetes:*:*:*:*:*:*:*:*",
+					VersionStartIncluding: "1.3.0",
+					VersionEndExcluding:   "1.3.9",
+				},
+				{
+					Cpe23Uri:              "cpe:2.3:a:kubernetes:kubernetes:*:*:*:*:*:*:*:*",
+					VersionStartIncluding: "1.4.0",
+					VersionEndExcluding:   "1.4.3",
+				},
+			},
+		},
+	}
+
+	tryUpdateCVEWithAffectedVersions(cve)
+
+	assert.EqualValues(t, expectedCPENodes, cve.Configurations.Nodes)
+}
